chat_processor/requirements.txt
requests
confluent-kafka
google-cloud-firestore
firebase-admin
google-generativeai
python-dotenv